% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_blds_by_grid.R
\name{calculate_blds_by_grid}
\alias{calculate_blds_by_grid}
\title{calculate total building footprint area within each hexagon of the grid}
\usage{
calculate_blds_by_grid(grid, blds)
}
\arguments{
\item{grid}{(sf object) output of radius_hex_grid}

\item{rds}{(sf object) building footprints sf object}
}
\value{
vector of total building area in meters that is the same length as the number of rows in the grid sf object
}
\description{
takes the result of radius_hex_grid and get_bldgs to sum the area of building footprints (not including height info) in meters within each hexagon of the grid
}
\examples{
\dontrun{
# calculate the summed building footprint area per each 1km hexagon in the grid
bld_ars<-uRbano::calculate_blds_by_grid(Rgrid, blds)
# add column for building area to the hex grid
Rgrid<-Rgrid \%>\% mutate(bld_ars=bld_ars)
}
}
