% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_buildings_by_grid.R
\name{calculate_buildings_by_grid}
\alias{calculate_buildings_by_grid}
\title{calculate total building footprint area within each hexagon of the grid}
\usage{
calculate_buildings_by_grid(grid, buildings)
}
\arguments{
\item{grid}{(sf object) output of create_hexagonal_grid}

\item{buildings}{(sf object) building footprints sf object}
}
\value{
vector of total building area in meters that is the same length as the number of rows in the grid sf object
}
\description{
takes the result of create_hexagonal_grid and get_buildings to sum the area of building footprints (not including height info) in meters within each hexagon of the grid
}
\examples{
# calculate the summed building footprint area per each 1km hexagon in the grid
# building_areas <- uRbano::calculate_buildings_by_grid(r_grid, buildings)
# add column for building area to the hex grid
# r_grid <- r_grid \%>\% dplyr::mutate(building_areas = building_areas)

}
