% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hexagonal_grid.R
\name{create_hexagonal_grid}
\alias{create_hexagonal_grid}
\title{Create a hexagonal grid over a given area}
\usage{
create_hexagonal_grid(area, hexagon_size)
}
\arguments{
\item{area}{An \code{sf} object representing the area to cover with the grid, typically a city radius.}

\item{hexagon_size}{The width of each hexagon in meters.}
}
\value{
An \code{sf} object of the hexagonal mesh. Hexagons intersecting the edge of the area are retained, not clipped.
}
\description{
This function uses \code{sf::st_make_grid} to produce a hexagonal mesh over a specified \code{sf} object. If the input is in WGS 84, it approximates the hexagon size from meters to decimal degrees.
}
\examples{
# Create a grid of 1km hexagons across a city_radius polygon
# hexagonal_grid <- uRbano::create_hexagonal_grid(city_radius, 1000)

}
