% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sampling_functions.R
\name{create_sampling_functions}
\alias{create_sampling_functions}
\title{create sampling functions}
\usage{
create_sampling_functions(
  hex_data,
  center_point = st_centroid(st_union(st_transform(hex_data, crs = 4326))),
  sample_size = 20,
  sample_type = NULL
)
}
\arguments{
\item{hex_data}{(sf object) output hex built density grid with a density index (dns_dx_) column that shows numeric levels of road and building density}

\item{center_point}{(sf object) single point sf object that defaults to the centroid of the union of the grid radius}

\item{sample_size}{(number) number of samples to generate, defaults to 10}

\item{sample_type}{can specify 5 type of sampling scenarios, defaults to a list with all 5}
}
\value{
line or polygon sf object
}
\description{
generates a set of helper functions that construct different sampling scenarios of the hex grid: sample types: cardinal_transects, concentric_rings, density_stratified, simple_random, or random_transect
}
\examples{

sampling_scenarios<-create_sampling_functions(grid)
# plot grid cells selected by concentric rings sampling
plot(sampling_scenarios[[2]])

}
