% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_and_trim_buildings.R
\name{fix_and_trim_buildings}
\alias{fix_and_trim_buildings}
\title{Trim building footprints and remove invalid geometries}
\usage{
fix_and_trim_buildings(buildings, radius)
}
\arguments{
\item{buildings}{A large \code{sfc} object of polygons, typically the output of \code{get_buildings}.}

\item{radius}{A polygon representing the extent to which the building footprints should be clipped, typically the output of \code{get_city_radius}.}
}
\value{
A large \code{sfc} object of polygons representing building footprints, with invalid geometries removed and trimmed to the specified extent.
}
\description{
This function is optional but recommended. It reduces the building footprint data to the extent of a given radius and removes invalid geometries to prevent errors in area calculations.
}
\examples{
# Fix invalid geometries and clip building footprints to the city radius
# buildings <- uRbano::fix_and_trim_buildings(buildings, city_radius)
}
