% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_n_trim_bldgs.R
\name{fix_n_trim_bldgs}
\alias{fix_n_trim_bldgs}
\title{trim building footprints to a smaller extent and remove invalid geometries}
\usage{
fix_n_trim_bldgs(blds, rad)
}
\arguments{
\item{blds}{(sfc) large simple features collection of polygons, output of get_bldgs}

\item{rad}{(polygon) output of get_city_rad or another extent to clip the building footprints to}
}
\value{
large simple features object of polygons representing building footprints in invalid geometries omitted and trimmed to the extent of the second parameter, rad
}
\description{
OPTIONAL BUT RECOMMENDED - reduce buildings to extent of a radius and filter out invalid geometries to avoid errors in calculating area
}
\examples{
\dontrun{
# fix invalid geometries and clip building footprints to city radius
blds<-uRbano::fix_n_trim_bldgs(blds, rad)
}
}
