% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_buildings.R
\name{get_buildings}
\alias{get_buildings}
\title{Get building footprints for a region of interest}
\usage{
get_buildings(region1, region2 = NULL)
}
\arguments{
\item{region1}{A character string of the primary region to look up in the online building footprint releases.}

\item{region2}{An optional character string for a second region.}
}
\value{
A large \code{sf} object of polygons representing building footprints. If \code{region2} is in a different country and the footprints cannot be combined, a list containing two \code{sf} objects is returned.
}
\description{
Extracts building footprint data from Microsoft's ML Building Footprints for one or two specified regions. This is useful for analyzing border cities. The data is released under an ODbL license.
}
\examples{
# Get building footprints for a US state
# buildings <- uRbano::get_buildings(region1 = "North Dakota")

# Get building footprints for two US states
# buildings <- uRbano::get_buildings(region1 = "North Dakota", region2 = "Minnesota")

# Get building footprints for a Canadian province
# buildings <- uRbano::get_buildings(region1 = "Manitoba")

# Get buildings for other countries or territories
# buildings <- uRbano::get_buildings(region1 = "PuertoRico")

# Get footprints for two regions in different countries
# buildings <- uRbano::get_buildings(region1 = "North Dakota", region2 = "Manitoba")
# nd_buildings <- buildings$region1
# mb_buildings <- buildings$region2
}
