% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-city-radius.R
\name{get_city_radius}
\alias{get_city_radius}
\title{Get city and create radius in meters}
\usage{
get_city_radius(city_name, radius)
}
\arguments{
\item{city_name}{(city_name) string of city name for selection from \code{maps::world.cities}
dataset in \code{maps}}

\item{radius}{(radius) integer of radius in meters}
}
\value{
(sfc) sf polygon simple feature collection of 1 feature with WGS 84 CRS
}
\description{
subsets user input city name from \code{maps::world.cities} dataset
in \code{maps} and creates a radius around the city center coordinates in specified
}
\examples{
# provide city name as a string and number of meters for radius from coordinates of city center
## 30km radius polygon for Minneapois USA
uRbano::get_city_radius(city_name = "Minneapolis", radius = 30000)

}
