% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_crs_to_meters.R
\name{transform_crs_to_meters}
\alias{transform_crs_to_meters}
\title{Transform a Coordinate Reference System (CRS) to a meters-based datum}
\usage{
transform_crs_to_meters(sf_object, method = "web_mercator")
}
\arguments{
\item{sf_object}{An \code{sf} object to be transformed.}

\item{method}{A string specifying the transformation method: "web_mercator" for EPSG:3857, or "NAD" for the appropriate NAD 83 UTM zone.}
}
\value{
An \code{sf} object with the modified CRS.
}
\description{
This function transforms the CRS of an \code{sf} object to a meters-based datum. It supports Web Mercator (EPSG:3857) and can automatically determine the appropriate NAD 83 UTM zone for features in North America.
}
\examples{
# Transform a city_radius object to the appropriate NAD 83 meters UTM zone
# city_radius_m <- uRbano::transform_crs_to_meters(city_radius, method = "NAD")

}
