% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_blds_by_grid.R, R/uRbano_functions.R
\name{calculate_blds_by_grid}
\alias{calculate_blds_by_grid}
\title{calculate total building footprint area within each hexagon of the grid}
\usage{
calculate_blds_by_grid(grid, blds)

calculate_blds_by_grid(grid, blds)
}
\arguments{
\item{grid}{An sf object of the hexagonal grid.}

\item{blds}{An sf object of the building footprints.}
}
\value{
vector of total building area in meters that is the same length as the number of rows in the grid sf object

A numeric vector of the building areas by grid cell.
}
\description{
takes the result of radius_hex_grid and get_bldgs to sum the area of building footprints (not including height info) in meters within each hexagon of the grid
}
\examples{
# calculate the summed building footprint area per each 1km hexagon in the grid
# bld_ars<-uRbano::calculate_blds_by_grid(Rgrid, blds)
# add column for building area to the hex grid
# Rgrid<-Rgrid \%>\% dplyr::mutate(bld_ars=bld_ars)

\dontrun{
rad <- get_city_rad("San Juan", 30000)
blds <- get_bldgs("PuertoRico")
blds <- fix_n_trim_bldgs(blds, rad)
Rgrid <- radius_hex_grid(rad, 1000)
# bld_ars <- calculate_blds_by_grid(Rgrid, blds)
}
}
