% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rds_by_grid.R, R/uRbano_functions.R
\name{calculate_rds_by_grid}
\alias{calculate_rds_by_grid}
\title{calculate total road length within each hexagon of the grid}
\usage{
calculate_rds_by_grid(grid, rds)

calculate_rds_by_grid(grid, rds)
}
\arguments{
\item{grid}{An sf object of the hexagonal grid.}

\item{rds}{An sf object of the roads.}
}
\value{
vector of total road length in meters that is the same length as the number of rows in the grid sf object

A numeric vector of the road lengths by grid cell.
}
\description{
takes the result of radius_hex_grid and extract_osm_rds to sum the length of road segments in meters within each hexagon of the grid
}
\examples{
# calculate the summed length of road segments per each 1km hexagon in the grid
# rd_lens<-uRbano::calculate_rds_by_grid(Rgrid, rds)
# add column for road length to the hex grid
# Rgrid<-Rgrid \%>\% mutate(rd_lens=rd_lens)

\dontrun{
rad <- get_city_rad("San Juan", 30000)
rds <- extract_osm_rds(rad)
Rgrid <- radius_hex_grid(rad, 1000)
rd_lens <- calculate_rds_by_grid(Rgrid, rds)
}
}
