% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_osm_roads.R
\name{extract_osm_roads}
\alias{extract_osm_roads}
\title{Get road features from the OSM Overpass API}
\usage{
extract_osm_roads(city_radius)
}
\arguments{
\item{city_radius}{An \code{sfc} object from which a bounding box can be extracted.}
}
\value{
An \code{sfc} object of LINESTRING features representing the road segments for the specified area.
}
\description{
This function queries the OpenStreetMap (OSM) Overpass API to retrieve road features within the bounding box of a given city radius.
}
\examples{
# Get the city radius (e.g., for Minneapolis)
city_radius <- uRbano::get_city_radius("Minneapolis", 30000)

# Extract OSM roads within the city's bounding box
roads <- uRbano::extract_osm_roads(city_radius = city_radius)

}
