% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_and_trim_buildings.R
\name{fix_and_trim_buildings}
\alias{fix_and_trim_buildings}
\title{trim building footprints to a smaller extent and remove invalid geometries}
\usage{
fix_and_trim_buildings(buildings, radius)
}
\arguments{
\item{buildings}{(sfc) large simple features collection of polygons, output of get_buildings}

\item{radius}{(polygon) output of get_city_radius or another extent to clip the
building footprints to}
}
\value{
large simple features object of polygons representing building
footprints in invalid geometries omitted and trimmed to the extent of the
second parameter, radius
}
\description{
OPTIONAL BUT RECOMMENDED - reduce buildings to extent of a
radius and dplyr::filter out invalid geometries to avoid errors in
calculating terra::area
}
\examples{
# fix invalid geometries and clip building footprints to city radius
# buildings<-uRbano::fix_and_trim_buildings(buildings, radius)
}
