% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_n_trim_bldgs.R, R/uRbano_functions.R
\name{fix_n_trim_bldgs}
\alias{fix_n_trim_bldgs}
\title{trim building footprints to a smaller extent and remove invalid geometries}
\usage{
fix_n_trim_bldgs(blds, rad)

fix_n_trim_bldgs(blds, rad)
}
\arguments{
\item{blds}{An sf object of building footprints.}

\item{rad}{An sf object of the city radius.}
}
\value{
large simple features object of polygons representing building footprints in invalid geometries omitted and trimmed to the extent of the second parameter, rad

An sf object of the trimmed and fixed building footprints.
}
\description{
OPTIONAL BUT RECOMMENDED - reduce buildings to extent of a radius and filter out invalid geometries to avoid errors in calculating area
}
\examples{
# fix invalid geometries and clip building footprints to city radius
# blds<-uRbano::fix_n_trim_bldgs(blds, rad)
\dontrun{
rad <- get_city_rad("San Juan", 30000)
blds <- get_bldgs("PuertoRico")
blds <- fix_n_trim_bldgs(blds, rad)
}
}
