% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-buildings.R
\name{get_buildings}
\alias{get_buildings}
\title{Get building footprints for a region of interest}
\usage{
get_buildings(region1, region2 = NULL)
}
\arguments{
\item{region1}{(string) character string of region to look up in online
building footprint releases}

\item{region2}{(string) optional second string}
}
\value{
large simple features object of polygons representing building
footprints. If region2 is in another country, a list containing two sf
objects is returned
}
\description{
takes user input as a string of the region or two regions -
for border cities - to extract building footprint data from Microsoft's
ML Building Footprints released under ODbL license:
https://minedbuildings.z5.web.core.windows.net.
}
\examples{
# get building footprints for US state
# buildings <- uRbano::get_buildings(region1="North Dakota")
# get building footprints for two US states
# buildings <- uRbano::get_buildings(region1="North Dakota", region2="Minnesota")
# get building footprints for Canadian province
# buildings <- uRbano::get_buildings(region1="Manitoba")
# get buildings for country or other territories
# buildings <- uRbano::get_buildings(region1="PuertoRico")
# two regions from different countries
# buildings <- uRbano::get_buildings(region1 = "North Dakota", region2="Manitoba")
# nd <- buildings$region1
# mb <- buildings$region2
}
