% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_city_rad.R, R/uRbano_functions.R
\name{get_city_rad}
\alias{get_city_rad}
\title{Get city and create radius in meters}
\usage{
get_city_rad(city, radius, state)

get_city_rad(city, radius, state)
}
\arguments{
\item{city}{A character string of the city name.}

\item{radius}{A numeric value of the radius in meters.}

\item{state}{An optional character string of the state name.}
}
\value{
(sfc) sf polgon simple feature collection of 1 feature with WGS 84 CRS

An sf object of the city radius.
}
\description{
subsets user input city name from \code{world.cities} dataset in \code{maps} and creates a radius around the city center coordinates in specified meters
}
\examples{
# provide city name as a string and number of meters for radius drawn from coordinates of city center
## 30km radius polygon for Minneapois USA
uRbano::get_city_rad(city="Minneapolis", radius=30000)


## if more than one city is found, a numbered list will be returned with information on city's country and population for the user to specify their selection by number

\dontrun{
rad <- get_city_rad("San Juan", 30000)
}
}
