% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_hex_grid.R, R/uRbano_functions.R
\name{radius_hex_grid}
\alias{radius_hex_grid}
\title{create hex grid of user-specified resolution across city radius}
\usage{
radius_hex_grid(radius, hex_size)

radius_hex_grid(radius, hex_size)
}
\arguments{
\item{radius}{An sf object of the city radius.}

\item{hex_size}{A numeric value of the hexagon size in meters.}
}
\value{
sf object of hexagon mesh with hexagons on the edge of the radius retained, not clipped

An sf object of the hexagonal grid.
}
\description{
uses sf::st_make_grid to produce an sf object of a hexagon mesh across the specified city radius. If given an sf object in WGS 84, it will approximate meters given for the width of the hexagons to decimal degrees
}
\examples{
# create grid of 1km hexagons across 'rad' polygon
# Rgrid<-uRbano::radius_hex_grid(rad, 1000)


\dontrun{
rad <- get_city_rad("San Juan", 30000)
Rgrid <- radius_hex_grid(rad, 1000)
}
}
