% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius-hex-grid.R
\name{radius_hex_grid}
\alias{radius_hex_grid}
\title{create hex grid of user-specified resolution across city radius}
\usage{
radius_hex_grid(radius, hex_size)
}
\arguments{
\item{radius}{(sf object) sf object, usually radius of a city from
get_city_radius, but could be another sf object}

\item{hex_size}{(numeric) number of meters across that each hex cell should be}
}
\value{
sf object of hexagon mesh with hexagons on the edge of the radius
retained, not clipped
}
\description{
uses sf::st_make_grid to produce an sf object of a hexagon mesh
across the specified city radius. If given an sf object in WGS 84, it will
approximate meters given for the width of the hexagons to decimal degrees
}
\examples{
# create grid of 1km hexagons across 'rad' polygon
# r_grid <- uRbano::radius_hex_grid(rad, 1000)

}
