% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_to_meters.R
\name{transform_to_meters}
\alias{transform_to_meters}
\title{transform coordinate reference system to meters datum}
\usage{
transform_to_meters(sf_data, method = "web_mercator")
}
\arguments{
\item{sf_data}{(sf object) sf object to transform}

\item{method}{(string) string that represents CRS transformation to perform: "web-mercator" for EPSG 3857, "NAD" for NAD 83}
}
\value{
sf object with modified CRS
}
\description{
Optional but recommended, transform buildings, roads and city radius features to meters datum with either web mercator EPSG 3857 or look up NAD 83 UTM zone to extract projection from the sf object parameter if spatial features are in North America
}
\examples{
# transform to NAD 83 meters UTM
# rad<- uRbano::transform_to_meters(rad, method="NAD")


}
