% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_to_meters.R, R/uRbano_functions.R
\name{transform_to_meters}
\alias{transform_to_meters}
\title{transform coordinate reference system to meters datum}
\usage{
transform_to_meters(sf_data, method = "utm")

transform_to_meters(sf_data, method = "utm")
}
\arguments{
\item{sf_data}{An sf object.}

\item{method}{A character string of the transformation method ("utm", "NAD", or "web_mercator").}
}
\value{
sf object with modified CRS

An sf object transformed to meters.
}
\description{
Optional but recommended, transform buildings, roads and city radius features to meters datum with either web mercator EPSG 3857 or look up NAD 83 UTM zone to extract projection from the sf object parameter if spatial features are in North America
}
\examples{
# transform to NAD 83 meters UTM
# rad<- uRbano::transform_to_meters(rad, method="NAD")


\dontrun{
rad <- get_city_rad("San Juan", 30000)
rad <- transform_to_meters(rad, method = "NAD")
}
}
